+++ src/Repositories/SqlStatementRepository.php
@@ -41,71 +41,77 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
      */
     private $statements = array(
         SqlStatementKeys::PRODUCT =>
-            'SELECT * FROM catalog_product_entity WHERE sku = :sku',
+            'SELECT * FROM ${table:catalog_product_entity} WHERE sku = :sku',
         SqlStatementKeys::PRODUCTS =>
-            'SELECT * FROM catalog_product_entity',
+            'SELECT * FROM ${table:catalog_product_entity}',
         SqlStatementKeys::PRODUCT_WEBSITE =>
-            'SELECT * FROM catalog_product_website WHERE product_id = :product_id AND website_id = :website_id',
+            'SELECT * FROM ${table:catalog_product_website} WHERE product_id = :product_id AND website_id = :website_id',
+        SqlStatementKeys::PRODUCT_WEBSITES_BY_SKU =>
+             'SELECT *
+                FROM ${table:catalog_product_website} t0
+          INNER JOIN ${table:catalog_product_entity} t1
+               WHERE t1.sku = :sku
+                 AND t0.product_id = t1.${pk:entity_id}',
         SqlStatementKeys::PRODUCT_DATETIMES =>
             'SELECT *
-               FROM catalog_product_entity_datetime
+               FROM ${table:catalog_product_entity_datetime}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::PRODUCT_DECIMALS =>
             'SELECT *
-               FROM catalog_product_entity_decimal
+               FROM ${table:catalog_product_entity_decimal}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::PRODUCT_INTS =>
             'SELECT *
-               FROM catalog_product_entity_int
+               FROM ${table:catalog_product_entity_int}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::PRODUCT_TEXTS =>
             'SELECT *
-               FROM catalog_product_entity_text
+               FROM ${table:catalog_product_entity_text}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::PRODUCT_VARCHARS =>
             'SELECT *
-               FROM catalog_product_entity_varchar
+               FROM ${table:catalog_product_entity_varchar}
               WHERE entity_id = :pk
                 AND store_id = :store_id',
         SqlStatementKeys::PRODUCT_VARCHAR_BY_ATTRIBUTE_CODE_AND_ENTITY_TYPE_ID_AND_STORE_ID =>
             'SELECT t1.*
-               FROM catalog_product_entity_varchar t1,
-                    eav_attribute t2
+               FROM ${table:catalog_product_entity_varchar} t1,
+                    ${table:eav_attribute} t2
               WHERE t2.attribute_code = :attribute_code
                 AND t2.entity_type_id = :entity_type_id
                 AND t1.attribute_id = t2.attribute_id
                 AND t1.store_id = :store_id',
         SqlStatementKeys::PRODUCT_VARCHAR_BY_ATTRIBUTE_CODE_AND_ENTITY_TYPE_ID_AND_STORE_ID_AND_VALUE =>
             'SELECT t1.*
-               FROM catalog_product_entity_varchar t1,
-                    eav_attribute t2
+               FROM ${table:catalog_product_entity_varchar} t1,
+                    ${table:eav_attribute} t2
               WHERE t2.attribute_code = :attribute_code
                 AND t2.entity_type_id = :entity_type_id
                 AND t1.attribute_id = t2.attribute_id
                 AND t1.store_id = :store_id
                 AND t1.value = BINARY :value',
         SqlStatementKeys::CATEGORY_PRODUCT =>
-            'SELECT * FROM catalog_category_product WHERE category_id = :category_id AND product_id = :product_id',
+            'SELECT * FROM ${table:catalog_category_product} WHERE category_id = :category_id AND product_id = :product_id',
         SqlStatementKeys::CATEGORY_PRODUCT_BY_SKU =>
             'SELECT t1.*
-               FROM catalog_category_product t1,
-                    catalog_product_entity t2
+               FROM ${table:catalog_category_product} t1,
+                    ${table:catalog_product_entity} t2
               WHERE t2.sku = :sku
                 AND t1.product_id = t2.entity_id',
         SqlStatementKeys::STOCK_ITEM =>
-            'SELECT * FROM cataloginventory_stock_item WHERE product_id = :product_id AND website_id = :website_id AND stock_id = :stock_id',
+            'SELECT * FROM ${table:cataloginventory_stock_item} WHERE product_id = :product_id AND website_id = :website_id AND stock_id = :stock_id',
         SqlStatementKeys::PRODUCT_RELATION =>
             'SELECT *
-                   FROM catalog_product_relation
-                  WHERE parent_id = :parent_id
-                    AND child_id = :child_id',
+               FROM ${table:catalog_product_relation}
+              WHERE parent_id = :parent_id
+                AND child_id = :child_id',
         SqlStatementKeys::CREATE_PRODUCT =>
             'INSERT
-               INTO catalog_product_entity
+               INTO ${table:catalog_product_entity}
                     (sku,
                      created_at,
                      updated_at,
@@ -121,7 +127,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :type_id,
                      :attribute_set_id)',
         SqlStatementKeys::UPDATE_PRODUCT =>
-            'UPDATE catalog_product_entity
+            'UPDATE ${table:catalog_product_entity}
                 SET sku = :sku,
                    created_at = :created_at,
                     updated_at = :updated_at,
@@ -132,14 +138,14 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE entity_id = :entity_id',
         SqlStatementKeys::CREATE_PRODUCT_WEBSITE =>
             'INSERT
-               INTO catalog_product_website
+               INTO ${table:catalog_product_website}
                     (product_id,
                      website_id)
              VALUES (:product_id,
                      :website_id)',
         SqlStatementKeys::CREATE_CATEGORY_PRODUCT =>
             'INSERT
-               INTO catalog_category_product
+               INTO ${table:catalog_category_product}
                     (category_id,
                      product_id,
                      position)
@@ -147,14 +153,14 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :product_id,
                      :position)',
         SqlStatementKeys::UPDATE_CATEGORY_PRODUCT =>
-            'UPDATE catalog_category_product
+            'UPDATE ${table:catalog_category_product}
                 SET category_id = :category_id,
                     product_id = :product_id,
                     position = :position
              WHERE  entity_id = :entity_id',
         SqlStatementKeys::CREATE_PRODUCT_DATETIME =>
             'INSERT
-               INTO catalog_product_entity_datetime
+               INTO ${table:catalog_product_entity_datetime}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -164,7 +170,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                     :store_id,
                     :value)',
         SqlStatementKeys::UPDATE_PRODUCT_DATETIME =>
-            'UPDATE catalog_product_entity_datetime
+            'UPDATE ${table:catalog_product_entity_datetime}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -172,11 +178,11 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::DELETE_PRODUCT_DATETIME =>
             'DELETE
-               FROM catalog_product_entity_datetime
+               FROM ${table:catalog_product_entity_datetime}
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_PRODUCT_DECIMAL =>
             'INSERT
-               INTO catalog_product_entity_decimal
+               INTO ${table:catalog_product_entity_decimal}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -186,7 +192,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                     :store_id,
                     :value)',
         SqlStatementKeys::UPDATE_PRODUCT_DECIMAL =>
-            'UPDATE catalog_product_entity_decimal
+            'UPDATE ${table:catalog_product_entity_decimal}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -194,11 +200,11 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::DELETE_PRODUCT_DECIMAL =>
             'DELETE
-               FROM catalog_product_entity_decimal
+               FROM ${table:catalog_product_entity_decimal}
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_PRODUCT_INT =>
             'INSERT
-               INTO catalog_product_entity_int
+               INTO ${table:catalog_product_entity_int}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -208,7 +214,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_PRODUCT_INT =>
-            'UPDATE catalog_product_entity_int
+            'UPDATE ${table:catalog_product_entity_int}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -216,11 +222,11 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::DELETE_PRODUCT_INT =>
             'DELETE
-               FROM catalog_product_entity_int
+               FROM ${table:catalog_product_entity_int}
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_PRODUCT_VARCHAR =>
             'INSERT
-               INTO catalog_product_entity_varchar
+               INTO ${table:catalog_product_entity_varchar}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -230,7 +236,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_PRODUCT_VARCHAR =>
-            'UPDATE catalog_product_entity_varchar
+            'UPDATE ${table:catalog_product_entity_varchar}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -238,11 +244,11 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::DELETE_PRODUCT_VARCHAR =>
             'DELETE
-               FROM catalog_product_entity_varchar
+               FROM ${table:catalog_product_entity_varchar}
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_PRODUCT_TEXT =>
             'INSERT
-               INTO catalog_product_entity_text
+               INTO ${table:catalog_product_entity_text}
                     (entity_id,
                      attribute_id,
                      store_id,
@@ -252,7 +258,7 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
                      :store_id,
                      :value)',
         SqlStatementKeys::UPDATE_PRODUCT_TEXT =>
-            'UPDATE catalog_product_entity_text
+            'UPDATE ${table:catalog_product_entity_text}
                 SET entity_id = :entity_id,
                     attribute_id = :attribute_id,
                     store_id = :store_id,
@@ -260,59 +266,64 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_PRODUCT_RELATION =>
             'INSERT
-                   INTO catalog_product_relation
+                   INTO ${table:catalog_product_relation}
                         (parent_id,
                          child_id)
                  VALUES (:parent_id,
                          :child_id)',
         SqlStatementKeys::DELETE_PRODUCT_TEXT =>
             'DELETE
-               FROM catalog_product_entity_text
+               FROM ${table:catalog_product_entity_text}
               WHERE value_id = :value_id',
         SqlStatementKeys::CREATE_STOCK_ITEM =>
-            'INSERT INTO cataloginventory_stock_item (%s) VALUES (:%s)',
+            'INSERT INTO ${table:cataloginventory_stock_item} (%s) VALUES (:%s)',
         SqlStatementKeys::UPDATE_STOCK_ITEM =>
-            'UPDATE cataloginventory_stock_item SET %s WHERE %s',
+            'UPDATE ${table:cataloginventory_stock_item} SET %s WHERE %s',
         SqlStatementKeys::DELETE_PRODUCT =>
             'DELETE
-               FROM catalog_product_entity
+               FROM ${table:catalog_product_entity}
               WHERE sku = :sku',
         SqlStatementKeys::DELETE_STOCK_ITEM_BY_SKU =>
-            'DELETE cataloginventory_stock_item
-               FROM cataloginventory_stock_item
-         INNER JOIN catalog_product_entity
-              WHERE catalog_product_entity.sku = :sku
-                AND cataloginventory_stock_item.product_id = catalog_product_entity.entity_id',
+            'DELETE t1.*
+               FROM ${table:cataloginventory_stock_item} t1
+         INNER JOIN ${table:catalog_product_entity} t2
+              WHERE t2.sku = :sku
+                AND t1.product_id = t2.entity_id',
         SqlStatementKeys::DELETE_PRODUCT_WEBSITE_BY_SKU =>
-            'DELETE catalog_product_website
-               FROM catalog_product_website
-         INNER JOIN catalog_product_entity
-              WHERE catalog_product_entity.sku = :sku
-                AND catalog_product_website.product_id = catalog_product_entity.entity_id',
+            'DELETE t1.*
+               FROM ${table:catalog_product_website} t1
+         INNER JOIN ${table:catalog_product_entity} t2
+              WHERE t2.sku = :sku
+                AND t1.product_id = t2.entity_id',
+        SqlStatementKeys::DELETE_PRODUCT_WEBSITE =>
+            'DELETE t1.*
+               FROM ${table:catalog_product_website} t1
+              WHERE t1.product_id = :product_id
+                AND t1.website_id = :website_id',
         SqlStatementKeys::DELETE_CATEGORY_PRODUCT =>
-            'DELETE
-               FROM catalog_category_product
+            'DELETE t1.*
+               FROM ${table:catalog_category_product} t1
               WHERE entity_id = :entity_id',
         SqlStatementKeys::DELETE_CATEGORY_PRODUCT_BY_SKU =>
-            'DELETE catalog_category_product
-               FROM catalog_category_product
-         INNER JOIN catalog_product_entity
-              WHERE catalog_product_entity.sku = :sku
-                AND catalog_category_product.product_id = catalog_product_entity.entity_id'
+            'DELETE t1.*
+               FROM ${table:catalog_category_product} t1
+         INNER JOIN ${table:catalog_product_entity} t2
+              WHERE t2.sku = :sku
+                AND t1.product_id = t2.entity_id'
     );
 
     /**
-     * Initialize the the SQL statements.
+     * Initializes the SQL statement repository with the primary key and table prefix utility.
+     *
+     * @param \IteratorAggregate<\TechDivision\Import\Utils\SqlCompilerInterface> $compilers The array with the compiler instances
      */
-    public function __construct()
+    public function __construct(\IteratorAggregate $compilers)
     {
 
-        // call the parent constructor
-        parent::__construct();
+        // pass primary key + table prefix utility to parent instance
+        parent::__construct($compilers);
 
-        // merge the class statements
-        foreach ($this->statements as $key => $statement) {
-            $this->preparedStatements[$key] = $statement;
-        }
+        // compile the SQL statements
+        $this->compile($this->statements);
     }
 }